Move ReadOnlyAdminMixin to core app
